requirements.txt
django
djangorestframework
pillow
